requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Pillow==8.3.0
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
pywin32==301
sqlparse==0.4.1
